feat: commandlock, add docs、unit tests，refactor namespace
