+++ netcorepal-cloud-framework.sln
@@ -141,6 +141,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetCorePal.Extensions.Dto",
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetCorePal.Extensions.Dto.Tests", "test\NetCorePal.Extensions.Dto.Tests\NetCorePal.Extensions.Dto.Tests.csproj", "{57FCB1FF-0A79-4172-B208-4AD7FC04EF26}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetCorePal.Extensions.Primitives.UnitTests", "test\NetCorePal.Extensions.Primitives.UnitTests\NetCorePal.Extensions.Primitives.UnitTests.csproj", "{F4C18BB0-F078-40AA-B09F-F2F4A06C2785}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -359,6 +361,10 @@ Global
 		{57FCB1FF-0A79-4172-B208-4AD7FC04EF26}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{57FCB1FF-0A79-4172-B208-4AD7FC04EF26}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{57FCB1FF-0A79-4172-B208-4AD7FC04EF26}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F4C18BB0-F078-40AA-B09F-F2F4A06C2785}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F4C18BB0-F078-40AA-B09F-F2F4A06C2785}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F4C18BB0-F078-40AA-B09F-F2F4A06C2785}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F4C18BB0-F078-40AA-B09F-F2F4A06C2785}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -422,6 +428,7 @@ Global
 		{59BB1378-ABEE-4A1E-BF14-7C807F83C8E3} = {605436C1-2560-47BA-99D4-7A80B628230F}
 		{02772177-21D1-4CF8-8F13-FBE4F5DE0F22} = {605436C1-2560-47BA-99D4-7A80B628230F}
 		{57FCB1FF-0A79-4172-B208-4AD7FC04EF26} = {DB5A8331-2B57-4F2B-8F7D-02E167BCD8BB}
+		{F4C18BB0-F078-40AA-B09F-F2F4A06C2785} = {DB5A8331-2B57-4F2B-8F7D-02E167BCD8BB}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D46201E5-1E3C-47A0-8584-4ACCC135CF9A}
