+++ src/Primitives/ICommandLock.cs
@@ -1,8 +1,9 @@
-using MediatR;
-using NetCorePal.Extensions.Primitives;
+namespace NetCorePal.Extensions.Primitives;
 
-namespace NetCorePal.Extensions.AspNetCore.CommandLocks;
-
+/// <summary>
+/// 表示一个命令锁
+/// </summary>
+/// <typeparam name="TCommand">要锁定的命令类型</typeparam>
 public interface ICommandLock<in TCommand> where TCommand : IBaseCommand
 {
     Task<CommandLockSettings> GetLockKeysAsync(TCommand command,
